+++ xdis/opcodes/opcode_21.py
@@ -21,11 +21,12 @@ This is similar to the opcode portion in Python 2.1's dis.py library.
 
 import xdis.opcodes.opcode_22 as opcode_22
 from xdis.opcodes.base import (
-    init_opdata,
+    extended_format_ATTR,
     extended_format_MAKE_FUNCTION_older,
     extended_format_RETURN_VALUE,
     finalize_opcodes,
     format_extended_arg,
+    init_opdata,
     rm_op,
     update_pj2,
 )
@@ -54,6 +55,8 @@ opcode_arg_fmt = {
 }
 
 opcode_extended_fmt = {
+    "LOAD_ATTR": extended_format_ATTR,
     "MAKE_FUNCTION": extended_format_MAKE_FUNCTION_older,
     "RETURN_VALUE": extended_format_RETURN_VALUE,
+    "STORE_ATTR": extended_format_ATTR,
 }