+++ xdis/opcodes/opcode_24.py
@@ -9,6 +9,7 @@ of stack usage, and opererand formatting functions.
 import xdis.opcodes.opcode_2x as opcode_2x
 from xdis.opcodes.base import (
     def_op,
+    extended_format_ATTR,
     extended_format_CALL_FUNCTION,
     extended_format_MAKE_FUNCTION_older,
     extended_format_RAISE_VARARGS_older,
@@ -52,7 +53,9 @@ opcode_arg_fmt = {
 
 opcode_extended_fmt = {
     "CALL_FUNCTION": extended_format_CALL_FUNCTION,
-    "RETURN_VALUE": extended_format_RETURN_VALUE,
-    "RAISE_VARARGS": extended_format_RAISE_VARARGS_older,
+    "LOAD_ATTR": extended_format_ATTR,
     "MAKE_FUNCTION": extended_format_MAKE_FUNCTION_older,
+    "RAISE_VARARGS": extended_format_RAISE_VARARGS_older,
+    "RETURN_VALUE": extended_format_RETURN_VALUE,
+    "STORE_ATTR": extended_format_ATTR,
 }