+++ xdis/opcodes/opcode_27pypy.py
@@ -9,6 +9,7 @@ of stack usage.
 import xdis.opcodes.opcode_27 as opcode_27
 from xdis.opcodes.base import (
     def_op,
+    extended_format_ATTR,
     extended_format_CALL_FUNCTION,
     extended_format_MAKE_FUNCTION_older,
     extended_format_RAISE_VARARGS_older,
@@ -70,7 +71,9 @@ finalize_opcodes(l)
 
 opcode_extended_fmt = {
     "CALL_FUNCTION": extended_format_CALL_FUNCTION,
+    "LOAD_ATTR": extended_format_ATTR,
+    "MAKE_FUNCTION": extended_format_MAKE_FUNCTION_older,
     "RAISE_VARARGS": extended_format_RAISE_VARARGS_older,
     "RETURN_VALUE": extended_format_RETURN_VALUE,
-    "MAKE_FUNCTION": extended_format_MAKE_FUNCTION_older,
+    "STORE_ATTR": extended_format_ATTR,
 }