+++ xdis/opcodes/opcode_36.py
@@ -22,6 +22,7 @@ of stack usage.
 
 from xdis.opcodes.base import(
     def_op,
+    extended_format_ATTR,
     extended_format_RAISE_VARARGS_older,
     extended_format_RETURN_VALUE,
     finalize_opcodes,
@@ -306,10 +307,12 @@ def extended_format_CALL_FUNCTION_KW(opc, instructions):
         return s
 
 opcode_extended_fmt = {
-    "CALL_METHOD": extended_format_CALL_METHOD,
     "CALL_FUNCTION": extended_format_CALL_FUNCTION,
     "CALL_FUNCTION_KW": extended_format_CALL_FUNCTION_KW,
+    "CALL_METHOD": extended_format_CALL_METHOD,
+    "LOAD_ATTR": extended_format_ATTR,
     "MAKE_FUNCTION": extended_format_MAKE_FUNCTION,
     "RAISE_VARARGS": extended_format_RAISE_VARARGS_older,
     "RETURN_VALUE": extended_format_RETURN_VALUE,
+    "STORE_ATTR": extended_format_ATTR,
 }